requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
oci
python-multipart